DatabaseFixed

+++ GHC2/Models/Admin.cs
@@ -7,12 +7,12 @@ namespace GHC2.Models
 {
     public class Admin
     {
-        public int Nid { get; set; }
+        public Int64 Nid { get; set; }
         public string Name { get; set; }
         public string Gender { get; set; }
         public DateTime? BitrhDate { get; set; }
         public string Address { get; set; }
-        public int? Phone { get; set; }
+        public Int64? Phone { get; set; }
         public string Email { get; set; }
         public string UserName { get; set; }
         public string Password { get; set; }
