+++ GHC2/Models/DiagnosePrescription.cs
@@ -7,8 +7,8 @@ namespace GHC2.Models
 {
     public class DiagnosePrescription
     {
-        public int PrescriptionId { get; set; }
-        public int DiagnoseId { get; set; }
+        public Int64 PrescriptionId { get; set; }
+        public Int64 DiagnoseId { get; set; }
 
         public virtual Diagnose Diagnose { get; set; }
     }
